Prune unneeded dependencies
This eliminates the dependency on the net.imagej:imagej gateway
in sc.iview.Main by constructing a SciJava context directly and
extracting needed services. Classpath elements desired only for running
Main are now:

* Marked optional to avoid infecting downstream consumers; and
* Runtime scope if not needed to compile the Main class.

My usual pattern for these sorts of testing classes is actually to put
them in src/test/java, but then you cannot run them with "mvn -Pexec"
which is unfortunate. So for now, we try this approach.