Fix the "crappy little window" bug
That weird little window that popped up was being created by the
UserInterface implementation (SwingSDIUI) when the Display was being
hooked to the DisplayViewer by the framework. The framework previously
assumed you'd always want the UserInterface to create a DisplayWindow
for you, which would then be fed to

    DisplayViewer#view(DisplayWindow, Display)

but SciView already has a window and would rather another window not be
created. So now we have

    DisplayViewer#view(UserInterface, Display)

which can be overridden by the SciViewDisplayViewer to skip that step.

This required a fix inside scijava-common; see

    scijava/scijava-common@8fc8deef25377fdb771273519021bdcbeb502b47