Fix Line command; let Box and Sphere give position
We now parse a String of three semicolon-separated values into a
Vector3. Some day it would be better to have a slick UI widget
for Vector3 inputs, but for now, String is good enough. We use
semicolon to avoid confusion with period (.) vs. comma (,) for
the decimal point in different localizations.