Update Travis CI configuration
This updates to the standard SciJava configuration.

No more macOS build. (We can restore it if desired.)

No more auto-upload to the update site, since we only want to do that
for releases now. Manually for the moment, but we could automate it.

No more building of topic branches -- only PRs and release branches.