Travis: check release.properties earlier
The release process results in release.properties no longer being
present. If we want to use that as an indicator for uploading to the
udpate site, then we should check it first and remember for later.

We also use sh instead of source now to trigger sciview_deploy.sh,
just in case Travis has an issue with using source.