POM: kick out additional repositories
Every repository here is a liability:

* Bad for performance: Maven has to check them sequentially for every
  needed artifact.
* Bad for reproducibility: if any of those go offline later, then all
  old builds referencing them will fail from then on.

To cut down on these issues, maven.scijava.org proxies all other
needed repositories.

In case the maven.scijava.org proxy is out of date: it is possible
to force the cache to expire immediately, and thus recheck next time,
on an artifact-by-artifact basis, as follows:

  maven-helper.sh invalidate-cache org.scijava:pom-scijava

Where "org.scijava:pom-scijava" is the artifact cache to expire.
The script lives at:

  https://github.com/scijava/scijava-scripts

It requires you have an account on maven.scijava.org with permission to
expire caches, and that you put your credentials into ~/.netrc like:

  machine maven.scijava.org
    login voltaire
    password zz123!DoubtIsAnUnpleasantConditionButCertaintyIsAbsurd

I use this, for example, after releasing a new pom-scijava to OSS
Sonatype, so that it gets picked up by maven.scijava.org immediately
instead of needing to wait up to 24 hours.