Disable VarsPane
The VarsPane is currently problematic because it is almost entirely filled with SciJava variables that simply confuse the user and are irrelevant to sciview. The VarsPane should be revisited and grounded in scenery variables, ideally using names from the scenegraph directly. It also should not have all of this extraneous stuff from the scijava context (that can be accessed through the scijava Context that is linked within the SciView object itself if necessary).