WIP - start tidying up the control settings dialog
The plan is to move the load/save persistence logic into
ControlsParameters, so that it's always aligned with the actual model
state. Then we can get rid of the PrefService and orig_* logic
completely in the NavigationControlsSettings dialog, in favor of just
getting the state from ControlsParameters upon dialog initialization,
and setting the ControlsParameters state any time widgets change.