fijiUpdate: do not forcibly downgrade things
The fijiUpdate task was using update-force-pristine which forcibly
resets all the JAR versions to those currently on the sciview update
site. So, if you do a populate, and then change something in sciview,
and then another populate... the second populate will cause many MB
worth of stuff to be redownloaded and overwrite your previous testing
version... only to then copy in a new testing version again. :face_palm:

So this commit switches from update-force-pristine to a normal update,
which will decline to overwrite the "locally modified" JARs.
The goal is to make the populate/test/edit loop less slow and wasteful.

This change could potentially result in a not-clean Fiji state from
which we then attempt to upload if, for example, you populate, then
remove a dependency, then populate again. In such a situation, the
Fiji.app would still have that old dependency as a "local only" file,
and then fijiUpload would subsequently upload it to the update site.